move statistics impl into .h
